+++ src/Phan/Analysis/NegatedConditionVisitor.php
@@ -30,7 +30,7 @@ use Phan\Language\UnionTypeBuilder;
  * TODO: Make $x > 0, $x < 0, $x >= 50, etc.  remove FalseType and NullType from $x
  * TODO: if (a || b || c || d) might get really slow, due to creating both ConditionVisitor and NegatedConditionVisitor
  */
-class NegatedConditionVisitor extends KindVisitorImplementation
+class NegatedConditionVisitor extends KindVisitorImplementation implements ConditionVisitorInterface
 {
     use ConditionVisitorUtil;